2017-08-09b

+++ code/lte-eric/agg_mgr.py
@@ -40,9 +40,10 @@ procDatetimeArr[1] = procDatetimeArr[1][:-3]
 ## globals
 prev_jobname = ""
 check_ctr = 0
+exportOnly = False
 
 # argv[1] - input dir
-# argv[2] - output dir
+# argv[2] - output dir - if empty, no export csv
 # argv[3] - cell lookup parquet dir
 # argv[4] - output parquet dir
 # argv[5] - option json string (optional); 
@@ -64,7 +65,9 @@ check_ctr = 0
 #              "logfile" : "", - empty = no log file
 #              "partitionNum": null, --> None in python (no coalesce)
 #              "overwrite": false/true, --> False/True in python
-#              "loadFactor": 3
+#              "loadFactor": 3,
+#              "exportHr": "3",
+#              "exportDaily": "N"
 #             }'
 # argv[6] - (optional) "cluster" or "client" mode
 
@@ -134,6 +137,10 @@ if 'loadFactor' not in optionJSON: # lte-eric: 3; umts-eric: 20
       optionJSON[u'loadFactor'] = 3
    else: # umts
       optionJSON[u'loadFactor'] = 20
+if 'exportHr' not in optionJSON:
+   optionJSON[u'exportHr'] = "3"
+if 'exportDaily' not in optionJSON:
+   optionJSON[u'exportDaily'] = "N"
 
 # init logger
 util.loggerSetup(__name__, optionJSON[u'logfile'], logging.DEBUG)
@@ -185,14 +192,19 @@ except OSError:
 # argv[1] - input dir
 input_dir = sys.argv[1]
 input_dir = input_dir.rstrip('/')
-if not os.path.isdir(input_dir):
+if input_dir == '':
+   exportOnly = True
+if input_dir != '' and not os.path.isdir(input_dir):
    util.logMessage("Failed to open input location \"%s\"!" % input_dir)
    util.logMessage("Process terminated.")
    util.endProcess(lockpath, 2)
 
 # create staging (if not exist)
-staging_dir = input_dir+'/staging'
-if not os.path.isdir(staging_dir): # create if not exist
+if exportOnly:
+   staging_dir = 'placeholder/staging'
+else:
+   staging_dir = input_dir+'/staging'
+if not exportOnly and not os.path.isdir(staging_dir): # create if not exist
    try:
       os.mkdir(staging_dir)
    except:
@@ -202,7 +214,7 @@ if not os.path.isdir(staging_dir): # create if not exist
 
 # create secondary staging
 staging_dir_sub = staging_dir + "/ttskpiagg_%s_%s_%s_%s_TMO" % (optionJSON[u'vendorFULL'], optionJSON[u'techUp'], procDatetimeArr[0], procDatetimeArr[1])
-if not os.path.isdir(staging_dir_sub): # create if not exist
+if not exportOnly and not os.path.isdir(staging_dir_sub): # create if not exist
    try:
       os.mkdir(staging_dir_sub)
    except:
@@ -213,7 +225,7 @@ if not os.path.isdir(staging_dir_sub): # create if not exist
 # argv[2] - output dir
 output_dir = sys.argv[2]
 output_dir = output_dir.rstrip('/')
-if not os.path.isdir(output_dir):
+if output_dir != '' and not os.path.isdir(output_dir):
    util.logMessage("Failed to open output location \"%s\"!" % output_dir)
    util.logMessage("Process terminated.")
    util.endProcess(lockpath, 2)
@@ -529,8 +541,8 @@ def canStartNewJob(statusJSON):
 def worker(seqfile):
 
 	global prev_jobname
-        seqfile_dir, seqfile_file = os.path.split(seqfile)
-	jobname = seqfile_file
+	seqfile_dir, seqfile_file = os.path.split(seqfile)
+	jobname = 'stg3_' + seqfile_file
 	jobname = jobname.replace(' ', '-') # for cluster mode, job name should not contain space - spark bug
 
 	util.logMessage("Task %s start..." % jobname)
@@ -586,29 +598,40 @@ def worker(seqfile):
 	exe_str_config_json = "'{\
 \"partitionNum\":%s, \
 \"overwrite\":%s, \
-\"loadFactor\":%s}'" % (pNum, ovrWr, ldFact)
+\"loadFactor\":%s, \
+\"exportHr\":\"%s\", \
+\"exportDaily\":\"%s\"}'" % (pNum, ovrWr, ldFact, optionJSON[u'exportHr'], optionJSON[u'exportDaily'])
 
 	# create python string
 	exec_str_py = "%s/kpi_agg_%s_%s.py" % (curr_py_dir, optionJSON[u'tech'], optionJSON[u'vendor'])
+	if exportOnly:
+		exec_str_input_dir = ""
+	else:
+		exec_str_input_dir = "%s/ttskpiraw_%s_%s_*_TMO.txt" % (seqfile, optionJSON[u'vendorFULL'], optionJSON[u'techUp'])
+	if output_dir == '':
+		exec_str_output_dir = ""
+	else:
+		exec_str_output_dir = "%s/ttskpiagg_%s_%s_TMO" % (output_dir, optionJSON[u'vendorFULL'], optionJSON[u'techUp'])
 	exec_str_app = "%s \
 \"%s\" \
-%s/ttskpiraw_%s_%s_\*_TMO.txt \
+\"%s\" \
 \"%s_%s_schema.json\" \
 \"%s\" \
 \"%s\" \
-\"%s/ttskpiagg_%s_%s_TMO\" \
+\"%s\" \
 %s \
 \"%s\"" % (exec_str_py, 
 		jobname, 
-		seqfile, optionJSON[u'vendorFULL'], optionJSON[u'techUp'], 
+		exec_str_input_dir, 
 		optionJSON[u'tech'], optionJSON[u'vendor'], 
 		input_celllookup_parq,
 		output_parq,
-		output_dir, optionJSON[u'vendorFULL'], optionJSON[u'techUp'], 
+		exec_str_output_dir, 
 		exe_str_config_json,
 		proc_mode)
 	if proc_mode != 'cluster': # client - support multi master (zookeeper)
-		exec_str_app += " &"
+		#exec_str_app += " &" # disable background job for agg to keep app locked and prevent next task from running if this one not finish
+		pass
 	else: # cluster - currently not support multi master (zookeeper)
 		pass
 
@@ -654,13 +677,42 @@ def main(input_dir, optionJSON):
    exit(0)
    '''
 
-   # go thru all seq file/folder
-   inputSeqPath = input_dir+"/ttskpiraw_%s_%s_*_TMO*.tgz" % (optionJSON[u'vendorFULL'], optionJSON[u'techUp'])
-   inputSeqList = glob.glob(inputSeqPath)
-   if len(inputSeqList) <= 0:  # no file
-      util.logMessage("No paser output to process: %s" % inputSeqPath)
-      os.system("rm -rf '%s'" % staging_dir_sub) # remove staging sub folder (since will not be removed by proc)
-      util.endProcess(lockpath, 0)
+   global exportOnly
+
+   if not exportOnly:
+      # go thru all seq file/folder
+      inputSeqPath = input_dir+"/ttskpiraw_%s_%s_*_TMO*.tgz" % (optionJSON[u'vendorFULL'], optionJSON[u'techUp'])
+      inputSeqList = glob.glob(inputSeqPath)
+      if len(inputSeqList) <= 0:  # no file
+         util.logMessage("No parser output to process: %s" % inputSeqPath)
+         os.system("rm -rf '%s'" % staging_dir_sub) # remove staging sub folder (since will not be removed by proc)
+         if output_dir == '': # if no input, and also no output, end process
+            util.endProcess(lockpath, 0)
+         else: # if no input, but have output, only do export
+            exportOnly = True
+
+
+   # export only mode
+   if exportOnly:
+      # submit one process to work on the whole folder (of multiple txt file)
+      try:
+         # get status
+         statusJSON = getStatusJSON_mesos()
+         bStartNewJob, delay_sec = canStartNewJob(statusJSON)
+         while (bStartNewJob == False):
+            time.sleep(delay_sec)
+            bStartNewJob, delay_sec = canStartNewJob(statusJSON) # retest after the sleep
+
+         # process file
+         worker(staging_dir_sub)
+
+      except Exception as e:
+         util.logMessage("Error: failed to export file %s\n%s" % (staging_dir_sub, e))
+      except:
+         util.logMessage("Unexpected error")
+
+      return 0
+
 
    # move seq file into staging_sub first to prevent other proc from touching them
    inputSeqStageList = []
